+++ types/transaction.go
@@ -6,8 +6,8 @@ import (
 )
 
 type Transaction struct {
-	RefBlockNum    int64      `json:"ref_block_num"`
-	RefBlockPrefix int64      `json:"ref_block_prefix"`
+	RefBlockNum    uint16     `json:"ref_block_num"`
+	RefBlockPrefix uint32     `json:"ref_block_prefix"`
 	Expiration     Time       `json:"expiration"`
 	Operations     Operations `json:"operations"`
 	Signatures     []string   `json:"signatures"`
